peripheral_pi/requirements.txt
pyserial>=3.5
RPi.GPIO>=0.7.1
smbus2>=0.4.3
Adafruit-Blinka>=8.0.0
adafruit-circuitpython-dht>=4.0.0
